Change smtp_response logic to M_parser
modified:   net/smtp/m_flow.h
modified:   net/smtp/m_flow_process.c
modified:   net/smtp/m_flow_tcp.c
modified:   net/smtp/m_flow_tcp_auth.c
modified:   net/smtp/m_flow_tcp_ehlo.c
modified:   net/smtp/m_flow_tcp_sendmsg.c
modified:   net/smtp/m_flow_tcp_smtp_response.c
modified:   net/smtp/m_flow_tcp_starttls.c

+++ net/smtp/m_flow_process.c
@@ -73,7 +73,7 @@ static M_state_machine_status_t M_state_write_chunk(void *data, M_uint64 *next)
 		return M_STATE_MACHINE_STATUS_NEXT;
 	}
 	next_chunk = &next_chunk[5];
-	chunk_len = next_chunk - slot->process.next_write_chunk;
+	chunk_len = (size_t)(next_chunk - slot->process.next_write_chunk);
 	M_buf_add_str_max(slot->out_buf, slot->process.next_write_chunk, chunk_len);
 	slot->process.next_write_chunk = next_chunk;
 	*next = STATE_WRITE_CHUNK_WAIT;
